+++ sql/personalized_feed.sql
@@ -69,7 +69,7 @@ SET search_path = public, extensions
 AS $$
 DECLARE
     user_embedding vector(384);
-    max_hot INT;
+    max_hot BIGINT;
 BEGIN
     -- Get user's interests embedding
     SELECT u.interests_embedding
@@ -84,10 +84,10 @@ BEGIN
 
     -- Get max hot score for normalization
     SELECT COALESCE(MAX(
-        COALESCE(j2.views, 0) * 6
-        + j2.cached_reaction_count * 3
-        + j2.cached_comment_count * 2
-        + j2.cached_bookmark_count * 2
+        COALESCE(j2.views, 0)::bigint * 6
+        + j2.cached_reaction_count::bigint * 3
+        + j2.cached_comment_count::bigint * 2
+        + j2.cached_bookmark_count::bigint * 2
     ), 1)
     INTO max_hot
     FROM public.journals j2
@@ -179,14 +179,14 @@ SECURITY DEFINER
 SET search_path = public, extensions
 AS $$
 DECLARE
-    max_hot INT;
+    max_hot BIGINT;
 BEGIN
     -- Get max hot score for normalization
     SELECT COALESCE(MAX(
-        COALESCE(j2.views, 0) * 6
-        + j2.cached_reaction_count * 3
-        + j2.cached_comment_count * 2
-        + j2.cached_bookmark_count * 2
+        COALESCE(j2.views, 0)::bigint * 6
+        + j2.cached_reaction_count::bigint * 3
+        + j2.cached_comment_count::bigint * 2
+        + j2.cached_bookmark_count::bigint * 2
     ), 1)
     INTO max_hot
     FROM public.journals j2
@@ -218,7 +218,7 @@ BEGIN
                 + j.cached_reaction_count * 3
                 + j.cached_comment_count * 2
                 + j.cached_bookmark_count * 2
-            )::INT AS h_score
+            )::BIGINT AS h_score
         FROM public.journals j
         LEFT JOIN public.users u ON u.id = j.user_id
         WHERE j.privacy = 'public'
